pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "build>=0.10.0"]
build-backend = "setuptools.build_meta"

[project]
name = "totoon"
version = "0.1.0"
description = "Convert any data format to TOON (Token-Oriented Object Notation)"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "totoon", email = "contact@totoon.dev"}
]
keywords = ["toon", "json", "llm", "data-format", "converter"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Markup",
]

dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
]
publish = [
    "build>=0.10.0",
    "twine>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/bug4fix/totoon"
Documentation = "https://github.com/bug4fix/totoon#readme"
Repository = "https://github.com/bug4fix/totoon"
Issues = "https://github.com/bug4fix/totoon/issues"

[tool.setuptools]
packages = ["totoon"]
[tool.setuptools.package-dir]
totoon = "python"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.ruff]
line-length = 100
target-version = "py38"

